tests/unit/requirements.txt
-r ../../requirements.txt
mock
jinja2
six
coverage
pyyaml
